+++ packages/web3-validator/src/types.ts
@@ -17,6 +17,7 @@ along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 
 import { AbiParameter } from 'web3-types';
 import { ZodIssueBase } from 'zod';
+import { JSONSchema4 } from 'json-schema';
 
 export type ValidInputTypes = Uint8Array | bigint | string | number | boolean;
 export type EthBaseTypes = 'bool' | 'bytes' | 'string' | 'uint' | 'int' | 'address' | 'tuple';
@@ -44,12 +45,7 @@ export type EthExtendedTypes =
 
 export type FullValidationSchema = ReadonlyArray<AbiParameter>;
 export type ShortValidationSchema = ReadonlyArray<
-	| string
-	| EthBaseTypes
-	| EthExtendedTypes
-	| EthBaseTypesWithMeta
-	| EthBaseTypesWithMeta
-	| ShortValidationSchema
+	string | EthBaseTypes | EthExtendedTypes | EthBaseTypesWithMeta | ShortValidationSchema
 >;
 export type ValidationSchemaInput = FullValidationSchema | ShortValidationSchema;
 
@@ -57,96 +53,16 @@ export type Web3ValidationOptions = {
 	readonly silent: boolean;
 };
 
-// is-my-json-valid types
 export type Json = string | number | boolean | Array<Json> | { [id: string]: Json };
 
-export type Schema = {
-	// version
-	$schema?: string;
-	$vocabulary?: string;
-	// pointers
-	id?: string;
-	$id?: string;
-	$anchor?: string;
-	$ref?: string;
-	definitions?: { [id: string]: Schema };
-	$defs?: { [id: string]: Schema };
-	$recursiveRef?: string;
-	$recursiveAnchor?: boolean;
-	// generic
-	type?: string | Array<string>;
-	required?: Array<string> | boolean;
-	default?: Json;
-	// constant values
-	enum?: Array<Json>;
-	const?: Json;
-	// logical checks
-	not?: Schema;
-	allOf?: Array<Schema>;
-	anyOf?: Array<Schema>;
-	oneOf?: Array<Schema>;
-	if?: Schema;
-	then?: Schema;
-	else?: Schema;
-	// numbers
-	maximum?: number;
-	minimum?: number;
-	exclusiveMaximum?: number | boolean;
-	exclusiveMinimum?: number | boolean;
-	multipleOf?: number;
-	divisibleBy?: number;
-	// arrays, basic
-	maxItems?: number;
-	minItems?: number;
-	additionalItems?: Schema;
-	// arrays, complex
-	contains?: Schema;
-	minContains?: number;
-	maxContains?: number;
-	uniqueItems?: boolean;
-	// strings
-	maxLength?: number;
-	minLength?: number;
-	format?: string;
-	pattern?: string;
-	// strings content
-	contentEncoding?: string;
-	contentMediaType?: string;
-	contentSchema?: Schema;
-	// objects
-	properties?: { [id: string]: Schema };
-	maxProperties?: number;
-	minProperties?: number;
-	additionalProperties?: Schema;
-	patternProperties?: { [pattern: string]: Schema };
-	propertyNames?: Schema;
-	dependencies?: { [id: string]: Array<string> | Schema };
-	dependentRequired?: { [id: string]: Array<string> };
-	dependentSchemas?: { [id: string]: Schema };
-	// see-through
-	unevaluatedProperties?: Schema;
-	unevaluatedItems?: Schema;
-	// Unused meta keywords not affecting validation (annotations and comments)
-	// https://json-schema.org/understanding-json-schema/reference/generic.html
-	// https://json-schema.org/draft/2019-09/json-schema-validation.html#rfc.section.9
-	title?: string;
-	description?: string;
-	deprecated?: boolean;
-	readOnly?: boolean;
-	writeOnly?: boolean;
-	examples?: Array<Json>;
-	$comment?: string;
-	// optimization hint and error filtering only, does not affect validation result
-	discriminator?: { propertyName: string; mapping?: { [value: string]: string } };
-	readonly eth?: string;
-	items?: Schema | Schema[];
-};
-
-type ValidationError = ZodIssueBase;
+export type ValidationError = ZodIssueBase;
 
 export interface Validate {
 	(value: Json): boolean;
 	errors?: ValidationError[];
 }
+type Schema = JSONSchema4 & { format?: string };
 
-export type JsonSchema = Schema;
+export type JsonSchema = Schema & { items?: Array<Schema> | Schema } & {
+	properties?: { [key: string]: Schema };
+};